Mati (#90)
* ## Backend  ### 1. ArticuloInsumoService.java  - **Método agregado**: `puedeVenderse(Long id)`   - Verifica si un insumo tiene stock suficiente (stockActual > 0)   - Ubicación: Al final de la clase, antes del cierre  ### 2. ArticuloInsumoController.java  - **Endpoint agregado**: `GET /api/insumos/{id}/can-be-sold`   - Retorna `boolean` indicando si el insumo puede venderse   - Ubicación: Junto a los otros métodos `@GetMapping`

* feat: Implementar CRUD completo de Unidades de Medida con baja lógica y validaciones avanzadas  ### Backend (Java Spring Boot):  **Modelo y DTO:** - Agregar campo `eliminado` a entidad UnidadMedida para baja lógica - Actualizar UnidadMedidaDto con campo eliminado y validaciones - Implementar soft delete manteniendo integridad referencial  **Repository:** - Métodos personalizados para encontrar solo unidades activas - Query para validación de duplicados case-insensitive - Soporte para operaciones con baja lógica  **Service:** - Validación de duplicados ignorando mayúsculas/minúsculas y espacios - Método `eliminarLogico()` para baja lógica - Método `restaurar()` para reactivar unidades eliminadas - Validaciones completas antes de crear/actualizar  **Controller:** - Endpoints `/api/unidades` (solo activas) y `/api/unidades/all` (todas) - Endpoints `/api/unidades/{id}/baja` y `/api/unidades/{id}/restaurar` - Manejo de errores con ResponseEntity y mensajes descriptivos

* feat: validación backend de denominaciones duplicadas para artículos manufacturados e insumos  **Backend ArticuloManufacturado:** - Endpoints de validación en ArticuloManufacturadoController - Métodos de validación en ArticuloManufacturadoService - Consultas de repository para verificar duplicados  **Backend ArticuloInsumo:** - Endpoints de validación en ArticuloInsumoController - Métodos de validación en ArticuloInsumoService - Consultas de repository para verificar duplicados  **Características implementadas:** - Validación case-insensitive con LOWER() - Endpoints GET /validate-denominacion y /validate-denominacion-all - Prevención de duplicados en save() y update() - Validación de elementos activos y eliminados - Manejo consistente de errores - Métodos transaccionales para integridad de datos - API RESTful para integración con frontend  Proporciona los endpoints necesarios para el sistema de validación del frontend y garantiza la integridad de datos a nivel de base de datos.

* feat: Implementar validación de restauración para categorías y artículos manufacturados  Backend: - Agregar endpoint GET /{id}/can-restore en CategoriaController - Agregar método canRestoreCategoria() en CategoriaService para validar jerarquía - Agregar endpoint GET /{id}/can-restore en ArticuloManufacturadoController   - Agregar método canRestoreArticulo() en ArticuloManufacturadoService - Validar que categorías padre no estén eliminadas antes de restaurar subcategorías - Validar que categorías no estén eliminadas antes de restaurar artículos - Validar jerarquía completa de categorías en restauración de artículos - Retornar mensajes de error específicos para cada caso de validación  Endpoints agregados: - GET /api/categorias/{id}/can-restore - GET /api/manufacturados/{id}/can-restore  Prevents: Restauración de elementos huérfanos y mantiene integridad referencial

* feat: Implementar validación de restauración para artículos insumo  Backend: - Agregar endpoint GET /{id}/can-restore en ArticuloInsumoController - Agregar método canRestoreArticuloInsumo() en ArticuloInsumoService para validar jerarquía - Modificar método restoreDeleted() con validación previa en ArticuloInsumoService - Validar que categorías no estén eliminadas antes de restaurar artículos insumo - Validar jerarquía completa de categorías en restauración de insumos - Retornar mensajes de error específicos para cada caso de validación - Mantener consistencia con validación de artículos manufacturados  Endpoints agregados: - GET /api/insumos/{id}/can-restore  Prevents: Restauración de insumos huérfanos y mantiene integridad referencial

* feat(backend): Migrar sistema de stock de Integer a Double para soporte decimal  MODELOS DE DOMINIO: - ArticuloInsumo: stockActual, stockMinimo (Integer → Double) - ArticuloManufacturadoDetalle: cantidad (Integer → Double) - DetallePedido: cantidad (Integer → Double) - PromocionDetalle: cantidadRequerida (Integer → Double) - CompraDetalle: cantidad (Integer → Double)  DTOs ACTUALIZADOS: - ProduccionAnalisisDTO: cantidadNecesaria, cantidadDisponible (Integer → Double) - CompraDetalleDTO: cantidad (Integer → Double) - ArticuloInsumoDTO: stockActual, stockMinimo (Integer → Double) - IngredienteDTO: cantidad (Integer → Double) - ArticuloManufacturadoDetalleDTO: cantidad (Integer → Double) - DetallePedidoDTO: cantidad (Integer → Double)  SERVICIOS MODIFICADOS: - ArticuloInsumoService:    * Validaciones de stock con Double   * Cálculos de disponibilidad con precisión decimal - StockService:   * Operaciones de suma/resta con valores decimales   * Validaciones de stock mínimo con Double - ArticuloValidacionService:   * Verificación de cantidades disponibles con decimales - HistorialService:   * Registro de movimientos con cantidades decimales - ProduccionService:   * Cálculo de ingredientes necesarios con precisión decimal  CONTROLADORES: - ArticuloController: Endpoints actualizados para recibir/devolver Double - CompraController: Manejo de cantidades decimales en compras - PedidoController: Procesamiento de cantidades decimales en pedidos  CONFIGURACIÓN DE DATOS: - DataInitializer:    * Stocks de insumos para elaborar: valores con .00 (ej: 10000.00)   * Stocks de productos terminados: valores exactos (ej: 300.00)   * Cantidades en recetas: precisión decimal (ej: 200.00g harina)   * Promociones corregidas con persistencia adecuada  VALIDACIONES MEJORADAS: - Verificación de stock con precisión decimal - Cálculos de costos con mayor exactitud - Validaciones de cantidades mínimas para compras - Control de stock crítico con decimales  BASE DE DATOS: - Migración de columnas INTEGER a DOUBLE/DECIMAL - Mantenimiento de precisión en operaciones aritméticas - Índices actualizados para nuevos tipos de datos  IMPACTO TÉCNICO: + Precisión en control de inventario (2.5kg vs 2kg) + Cálculos exactos en recetas y costos + Soporte para unidades fraccionarias (litros, kilos) + Eliminación de errores de redondeo en operaciones + Base sólida para reportes financieros precisos  COMPATIBILIDAD: - Migración backward-compatible - Datos existentes preservados con conversión automática - APIs mantienen contratos con nuevos tipos

* 📦 Feat (unidad medida): Agregar inicializador y métodos de eliminación lógica y física

- Crear UnidadMedidaInitializer para cargar unidades de medida básicas al iniciar la aplicación
- Implementar eliminación lógica por defecto en UnidadMedidaController
- Agregar método deletePhysically para eliminación física en UnidadMedidaController
- Incluir método findByIdIncludingDeleted en UnidadMedidaRepository para buscar unidades eliminadas
- Modificar UnidadMedidaService para manejar eliminación lógica y física

---------

+++ src/main/java/utn/saborcito/El_saborcito_back/dto/ArticuloManufacturadoDetalleDTO.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @Builder
 public class ArticuloManufacturadoDetalleDTO {
     private Long id;
-    private Integer cantidad;
+    private Double cantidad;    // ✅ Cambio de Integer a Double
     private ArticuloInsumoDTO articuloInsumo; // Usar ArticuloInsumoDTO para el insumo
     // private Long articuloManufacturadoId; // Opcional: si se necesita el ID del
     // manufacturado padre
